move lower_hybrid_frequency to plasmapy.formulary.frequencies
